Moved authentication methods to a separate file
