+++ src/Components/ManageUsers/CreateOrUpdateUser.js
@@ -3,7 +3,8 @@ import React, { useState, useContext, useEffect } from 'react';
 import { Button, FormControl, InputLabel, MenuItem, Select, TextField } from '@mui/material';
 import Grid from '@mui/material/Grid';
 import ManageUsersContext from "./ManageUsersContext";
-import { userRegister, updateUser } from "../../helpers/users";
+import { updateUser } from "../../helpers/users";
+import { userRegister } from "../../helpers/authentication";
 
 const CreateOrUpdateUser = () => {
     const { control, handleSubmit, reset, formState: { errors, isSubmitted, isSubmitSuccessful, isValid }, watch, setError, clearErrors, setValue, getValues }
